fixing warnings from the headers file.

+++ include/pomar/MorphologicalTree/MorphologicalTree.hpp
@@ -73,7 +73,7 @@ namespace pomar
     void _reconstructNode(int id, std::vector<int>& rec);
   protected:
     std::vector<MTNode<T>> _nodes;
-    std::vector<size_t> _cmap;
+    std::vector<int> _cmap;
   };
 
   /* ========================= MORPHOLOGICAL TREE - TRANSVERSAL ================================ */
@@ -133,7 +133,7 @@ namespace pomar
       parentNode.addChild(node.id());
     }
 
-    for (int i = 0; i < vertices.size(); i++) {
+    for (size_t i = 0; i < vertices.size(); i++) {
       if (_cmap[i] == UNDEF) {
 	_cmap[i] = _cmap[parent[i]];
 	_nodes[_cmap[i]].addVertexIndex(i);
